Enable DirectPred model to work with multiple target variables.
It will create a MLP head for each. It will distinguish the data types (numerical/categorical) and compute a regression/classificaiton accordingly.
Also the loss will be accordingly computed. Samples with missing labels will not be thrown out.
They will be kept in the loop, but only ignored during computing the loss values.